refactor category details screen ui by separating the content into different components
