+++ app/src/main/res/values/strings.xml
@@ -104,4 +104,5 @@
     <string name="category_deleted_successfully">Category deleted successfully</string>
     <string name="an_error_occured_while_deleting_the_category">An error occured while deleting the category</string>
     <string name="an_error_occurred_while_updating_the_category">An error occurred while updating the category</string>
+    <string name="category_updated_successfully">Category updated successfully</string>
 </resources>
